requirements.txt
SQLAlchemy~=2.0.43
python-decouple~=3.8
aiohttp~=3.12.15
aiogram~=3.22.0